requirements.txt
# requirements.txt
kafka-python>=2.0.2
pyspark>=3.3.0
